move taxonomy helpers to a separate file
