+++ frontend/src/layout/navigation-3000/sidebars/dataManagement.ts
@@ -3,7 +3,7 @@ import { loaders } from 'kea-loaders'
 import { subscriptions } from 'kea-subscriptions'
 import api from 'lib/api'
 import { TaxonomicFilterGroupType } from 'lib/components/TaxonomicFilter/types'
-import { getFilterLabel } from 'lib/taxonomy'
+import { getFilterLabel } from '~/taxonomy/helpers'
 import { actionsFuse, actionsLogic } from 'scenes/actions/actionsLogic'
 import { sceneLogic } from 'scenes/sceneLogic'
 import { Scene } from 'scenes/sceneTypes'
