+++ frontend/src/lib/components/DefinitionPopover/DefinitionPopover.tsx
@@ -8,7 +8,7 @@ import { TaxonomicFilterGroupType } from 'lib/components/TaxonomicFilter/types'
 import { dayjs } from 'lib/dayjs'
 import { LemonMarkdown } from 'lib/lemon-ui/LemonMarkdown'
 import { Link } from 'lib/lemon-ui/Link'
-import { getCoreFilterDefinition } from 'lib/taxonomy'
+import { getCoreFilterDefinition } from '~/taxonomy/helpers'
 import { eventUsageLogic } from 'lib/utils/eventUsageLogic'
 
 import { CoreFilterDefinition, UserBasicType } from '~/types'
