+++ frontend/src/lib/components/PropertyKeyInfo.tsx
@@ -3,7 +3,7 @@ import './PropertyKeyInfo.scss'
 import { LemonDivider, TooltipProps } from '@posthog/lemon-ui'
 import clsx from 'clsx'
 import { Popover } from 'lib/lemon-ui/Popover'
-import { getCoreFilterDefinition, PropertyKey } from 'lib/taxonomy'
+import { getCoreFilterDefinition, PropertyKey } from '~/taxonomy/helpers'
 import React, { useState } from 'react'
 
 import { TaxonomicFilterGroupType } from './TaxonomicFilter/types'
