+++ frontend/src/scenes/data-management/definition/DefinitionEdit.tsx
@@ -10,7 +10,7 @@ import { LemonButton } from 'lib/lemon-ui/LemonButton'
 import { LemonField } from 'lib/lemon-ui/LemonField'
 import { LemonSelect } from 'lib/lemon-ui/LemonSelect'
 import { LemonTextArea } from 'lib/lemon-ui/LemonTextArea/LemonTextArea'
-import { getFilterLabel, isCoreFilter } from 'lib/taxonomy'
+import { getFilterLabel, isCoreFilter } from '~/taxonomy/helpers'
 import { definitionEditLogic } from 'scenes/data-management/definition/definitionEditLogic'
 import { definitionLogic, DefinitionLogicProps } from 'scenes/data-management/definition/definitionLogic'
 import { SceneExport } from 'scenes/sceneTypes'
