+++ frontend/src/scenes/session-recordings/filters/ReplayTaxonomicFilters.tsx
@@ -5,7 +5,7 @@ import { PropertyKeyInfo } from 'lib/components/PropertyKeyInfo'
 import { TaxonomicFilter } from 'lib/components/TaxonomicFilter/TaxonomicFilter'
 import { TaxonomicFilterGroupType, TaxonomicFilterValue } from 'lib/components/TaxonomicFilter/types'
 import { universalFiltersLogic } from 'lib/components/UniversalFilters/universalFiltersLogic'
-import { getFilterLabel } from 'lib/taxonomy'
+import { getFilterLabel } from '~/taxonomy/helpers'
 import { useState } from 'react'
 
 import { PropertyFilterType } from '~/types'
