+++ frontend/src/scenes/session-recordings/player/inspector/components/SimpleKeyValueList.tsx
@@ -2,7 +2,7 @@
 
 import { PropertyKeyInfo } from 'lib/components/PropertyKeyInfo'
 import { TaxonomicFilterGroupType } from 'lib/components/TaxonomicFilter/types'
-import { getCoreFilterDefinition } from 'lib/taxonomy'
+import { getCoreFilterDefinition } from '~/taxonomy/helpers'
 import { ReactNode, useEffect, useState } from 'react'
 
 export interface SimpleKeyValueListProps {
