+++ frontend/src/scenes/surveys/SurveyEdit.tsx
@@ -29,7 +29,7 @@ import { LemonField } from 'lib/lemon-ui/LemonField'
 import { LemonRadio, LemonRadioOption } from 'lib/lemon-ui/LemonRadio'
 import { Tooltip } from 'lib/lemon-ui/Tooltip'
 import { featureFlagLogic as enabledFeaturesLogic } from 'lib/logic/featureFlagLogic'
-import { getPropertyKey } from 'lib/taxonomy'
+import { getPropertyKey } from '~/taxonomy/helpers'
 import { formatDate } from 'lib/utils'
 import { useState } from 'react'
 import { featureFlagLogic } from 'scenes/feature-flags/featureFlagLogic'
